Rework driver initialization / teardown to please the "usual" use case.
The driver now gets initialized when the first (and usually the only) channel
gets opened. It gets shut down after this very channel has been closed.

NOTE: This is optimized for the one-channel usecase, which makes most sense
for my use case. Please holler, if it breaks yours, then we need to reevaluate.
In the meantime this improves the noise induced due to the unconditional
initialization / deinitialization by the means of the library
`__constructor__` and `__destructor__` attributes.