Rework driver initialization / teardown again.
The auto-initialize part when opening the first channel is quite helpful,
but it turns out the auto-finalizer when closing the last channel does more
harm than good.

There this commit splits the OPTION_CANAPI_TOUCAN_DYLIB which is controlling
the auto init / auto finalize via the means of __constructor__ / __destructor__
into two disjunct options:
 - OPTION_CANAPI_TOUCAN_AUTO_INIT
 - OPTION_CANAPI_TOUCAN_AUTO_FINALIZE

The Swift package turns on the auto finalizer, but leaves the auto init off.